requirements.txt
pymupdf>=1.23.8
pillow>=10.0.0
reportlab>=4.0.4
flask>=3.0.0
ebooklib>=0.18
lxml>=4.9.0
